+++ src/core/computed/make-cu-ob-state.js
@@ -1,0 +1,13 @@
+
+export default function (state, depKeys) {
+  return new Proxy(state, {
+    get: function (target, key) {
+      if (!depKeys.includes(key)) depKeys.push(key);
+      return target[key];
+    },
+    set: function (target, key) {
+      target[key] = target[key];
+      return true;
+    }
+  });
+}